pyproject.toml
[tool.poetry]
name = "food-classifier"
version = "0.1.0"
description = "Classifier for Food101 dataset using PyTorch (torchvision.datasets.Food101)"
authors = ["chrrei <christian.reinecke@protonmail.com>"]
readme = "README.md"
packages = [
    {include = "models"},
    {include = "datamodules"},
    {include = "trainer"},
    {include = "experiments"}
]

[tool.poetry.dependencies]
python = "^3.11"
torchvision = "^0.15.2"
matplotlib = "^3.7.1"
pytorch-lightning = "^2.0.3"
torch = "^2.0.1"
pyyaml = "^6.0"
numpy = "^1.24.3"
torchmetrics = "^0.11.4"
tensorboard = "^2.13.0"

[tool.poetry.scripts]
main = "main:main"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
